requirements.txt
psycopg2-binary
SQLAlchemy
pandas
python-dotenv
matplotlib
seaborn
plotly
